integrating the username fetching into Header and removing it from Menu and SingleMatch
