+++ frontend/src/components/Menu.tsx
@@ -1,5 +1,4 @@
-import React, { useState, useEffect } from 'react';
-import Header from './Header'; 
+import React from 'react';
 import { useTranslation } from 'react-i18next'
 import SettingsMenu from './SettingsMenu'
 import { FaPlay, FaTrophy, FaChartBar } from 'react-icons/fa'
@@ -32,50 +31,8 @@ const MenuCard = ({ title, colorClass, icon, to }: MenuCardProps) => (
 // ---------------------------------------
 
 const Menu = () => {
-
-   const handleLogout = () => {
-    console.log('Logging out from Menu component...');
-  };
-
-  const handleFriendsClick = () => {
-    console.log('Navigating to Friends list...');
-  };
    
   const { t } = useTranslation();
-
-  const [userName, setUserName] = useState('Loading...');
-  const SelfAPI = 'http://localhost:3000/api/user/profile/self';
-
-  useEffect(() => {
-		const fetchUserName = async () => {
-			try {
-				const response = await fetch(SelfAPI, {
-					method: 'GET',
-					credentials: 'include',
-					headers: {
-					},
-				});
-				if (response.status === 401) {
-        			throw new Error("User not authenticated.");
-    			}
-				if (!response.ok) {
-					throw new Error(`HTTP Error: ${response.status}`);
-				}
-				const result = await response.json();
-
-				if (result.data && result.data.userName) {
-					//console.log(result.data.userName);
-					setUserName(result.data.userName);
-				} else {
-					throw new Error(result.message);
-				}
-			} catch (e) {
-				console.error("Failed to fetch user profile: ", e);
-				setUserName("Player1");
-			}
-		};
-		fetchUserName();
-  }, []);
 
   return (
     <div className="relative h-screen overflow-hidden">
