+++ frontend/src/components/SingleMatch.tsx
@@ -1,4 +1,4 @@
-import React, { useState, useEffect } from 'react';
+import React from 'react';
 import Header from './Header'; 
 import { useTranslation } from 'react-i18next'
 import SettingsMenu from './SettingsMenu'
@@ -16,60 +16,18 @@ const colorClasses = {
 
 const SingleMatch = () => {
 
-  const handleLogout = () => {
-	console.log('Logging out from Menu component...');
-  };
-
-  const handleFriendsClick = () => {
-	console.log('Navigating to Friends list...');
-  };
-
   const { t } = useTranslation();
-  
-  const [userName, setUserName] = useState('Loading...');
-  const SelfAPI = 'http://localhost:3000/api/user/profile/self';
-
-  useEffect(() => {
-	const fetchUserName = async () => {
-		try {
-			const response = await fetch(SelfAPI, {
-				method: 'GET',
-				credentials: 'include',
-				headers: {
-				},
-			});
-			if (response.status === 401) {
-				throw new Error("User not authenticated.");
-			}
-			if (!response.ok) {
-				throw new Error(`HTTP Error: ${response.status}`);
-			}
-			const result = await response.json();
-
-  			if (result.data && result.data.userName) {
-				console.log(result.data.userName);
-				setUserName(result.data.userName);
-			} else {
-				throw new Error(result.message);
-			}
-		} catch (e) {
-			console.error("Failed to fetch user profile: ", e);
-			setUserName("Player1");
-		}
-	};
-	fetchUserName();
-  }, []);
 
   return (
 	<div className="relative h-screen overflow-hidden">
 		
 	  <SettingsMenu />
-	  <Header
+	  {/* <Header
 		appName="PONG"
 		playerName={userName}
 		onFriendsClick={handleFriendsClick}
 		onLogout={handleLogout}
-	  />
+	  /> */}
 	  <div 
 		className={`${colorClasses.bgGlow} flex items-center justify-center p-4`}
 		style={{ height: 'calc(100vh - 3.5rem)' }} // make sure content fits below header
